// frontend/lib/generated/fastapi/schemas/leaderboardPeriod.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */

export type LeaderboardPeriod = typeof LeaderboardPeriod[keyof typeof LeaderboardPeriod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LeaderboardPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;